Major changes in the files structure & names
everything should work as before.

Some files moved between directories. Other changed names. Some files were
split to more than one (server.py).

The only code-related change is yapapi_executor_config on app (instead
of get_config() function).
