--- requestor/erigon_services/__init__.py
@@ -1,2 +1,0 @@
-from .erigon import Erigon
-from .pseudo_erigon import PseudoErigon
